Guess my number
The game is called "Guess my number".

The goal is to guess the number (between 1 and 20) which is hidden behind the question box.

While trying to guess, the system tells you whether it is too high or too low and points out how many tries (out of 20) you have, as well as the highest score that you've achieved.